Inline server
