--- rs/ic_os/vsock/src/host/mod.rs
@@ -1,4 +1,0 @@
-mod agent;
-mod command_utilities;
-mod hsm;
-pub(crate) mod server;
