+++ rs/ic_os/vsock/src/lib.rs
@@ -4,8 +4,8 @@ pub use client::LinuxVSockClient;
 pub use client::{MockVSockClient, VSockClient};
 
 #[cfg(target_os = "linux")]
-mod host;
+mod server;
 #[cfg(target_os = "linux")]
-pub use host::server::run_server;
+pub use server::run_server;
 
 pub mod protocol;
